zivis-sim-py/requirements.txt
fastapi
uvicorn
redis
pgvector
psycopg2-binary
openai>=1.0.0
tiktoken
langchain>=0.1.16
langchain-community>=0.0.18
datasets
faker
python-dotenv
